Added CombatOps seam for combat modularization.
